Revert "change callback to uint32"
This reverts commit 34a9b9afea7d506a6c47efcdce101dc172f69e74.

+++ tasmota/tasmota_xdrv_driver/xdrv_01_9_webserver.ino
@@ -3696,7 +3696,7 @@ void CmndCors(void)
  * Interface
 \*********************************************************************************************/
 
-bool Xdrv01(uint32_t function)
+bool Xdrv01(uint8_t function)
 {
   bool result = false;
