+++ tasmota/tasmota_xsns_sensor/xsns_08_htu21.ino
@@ -255,7 +255,7 @@ void HtuShow(bool json)
  * Interface
 \*********************************************************************************************/
 
-bool Xsns08(uint32_t function)
+bool Xsns08(uint8_t function)
 {
   if (!I2cEnabled(XI2C_09)) { return false; }
 
@@ -284,3 +284,4 @@ bool Xsns08(uint32_t function)
 
 #endif  // USE_HTU
 #endif  // USE_I2C
+
