+++ tasmota/tasmota_xsns_sensor/xsns_40_pn532.ino
@@ -544,7 +544,7 @@ void PN532_Show(void) {
  * Interface
 \*********************************************************************************************/
 
-bool Xsns40(uint32_t function) {
+bool Xsns40(uint8_t function) {
   bool result = false;
 
   if (FUNC_INIT == function) {
