+++ tasmota/tasmota_xsns_sensor/xsns_64_hrxl.ino
@@ -98,7 +98,7 @@ void HRXLShow(bool json)
  * Interface
 \*********************************************************************************************/
 
-bool Xsns64(uint32_t function)
+bool Xsns64(uint8_t function)
 {
    if (!PinUsed(GPIO_HRXL_RX)) { return false; }
