+++ tasmota/tasmota_xsns_sensor/xsns_83_neopool.ino
@@ -2534,7 +2534,7 @@ void CmndNeopoolIONRes(void)
  * Interface
 \*********************************************************************************************/
 
-bool Xsns83(uint32_t function)
+bool Xsns83(uint8_t function)
 {
   bool result = false;
